feat(edge): encapsulate edgestore api to a config file
